Update D-Bus proxy for UPnP server when reported again.
This partially fixes the problem with nameless UPnP devices.

The commit works for the case when dleyna-server-service has crashed,
and then the new instance reports all device again. Since we haven't
seen any LostServer signals from the crashed instance, our server list
entries still referred to the crashed instance. All we did in previous
versions was to consider this as an oddity and to emit a bug message
to the log. Now we update the D-Bus proxy for all "duplicate" server
objects.

Note that this is only a partial fix. It only works if and when dLeyna
reports new servers, after dLeyna has been restarted, and only for
servers that are actually reported by the new dLeyna instance. Another
fix is required which monitors dLeyna vitality, and clears the list
when dLeyna disappears (this will, in fact, make this commit
superfluous, but we'll still keep it in place as an act of defensive
programming).

Change-Id: Ibc30b6e7c0499b17e3ff6c1bfa050349ab144d78