Watch dLeyna on D-Bus, reconnect if it disappears.
This commit accomplishes one simple thing: when dLeyna disappears on
D-Bus (possibly due to a crash), then the UPnP server list is wiped
out (tested by killing dleyna-server-service).

Of course, the association with the disappeared dLeyna process must be
removed as well in this case, and a new connection with dLeyna must be
triggered (counting on D-Bus service activation). This change also
means that the D-Bus proxy object for dLeyna may be temporary NULL or
in some unconnected state, a situation the code wasn't prepared for.
It should work now, but there might be some new bugs lurking now in
the old code.

Finally, no more "Nameless UPnP device" entries.

Change-Id: I92d0336ccf6fe43aea143a848c5889de4ec7c803