UPnP: Avoid crash on D-Bus error (#1512).
We catch the exception thrown when accessing an invalid list item,
decode the original GLib D-Bus error which led to the invalid item,
and do our best to handle the situation. The program doesn't crash
anymore, but the results look weird. Seems like the UPnP server or
dLeyna is buggy.

Change-Id: Iceadc90f8b6b1af0b3de08a91eea7db0fef5b170