meson: Fix path to D-Bus submodule.
Before, the build system could end up picking the wrong path to the
D-Bus submodule if it is found at the given relative path outside of
the project root. In case that accidentally picked up submodule is
different from the one used by the project, the weirdest of errors
were triggered.

The problem is the use of a plain strings for a path so that Meson
couldn't really make sense of it. Now Meson is aware, and the problem
does not exist anymore.

Change-Id: I49b9c7942e65b0a5b3c41e34bf343c7e62aa2685